hook up component class
